+++ src/App.jsx
@@ -30,12 +30,7 @@ function App() {
 
   return (
     <>
-      <Navbar 
-        goToCourseList={goToCourseList}
-        toggleOverlay={toggleOverlay}
-        setPage={setPage}
-        />
-      {overlayOpen && <ProfileOverlay userName="Guest" setPage={(p) => { setOverlayOpen(false); setPage(p) }} />}
+      <Navbar />
       {page === 'Homepage' && (
         <Homepage toggleOverlay={toggleOverlay} goToCourseList={goToCourseList} />
       )}
@@ -66,6 +61,7 @@ function App() {
         </>
       )}
       {page === 'error' && <Error />}
+      {overlayOpen && <ProfileOverlay userName="Guest" setPage={(p) => { setOverlayOpen(false); setPage(p) }} />}
     </>
   )
 }
