+++ src/CourseProfile.jsx
@@ -41,6 +41,7 @@ function CourseProfile({courseName, courseDesc, prereqs, subject, difficulty, hw
                 <div id="whyCantIGethrToWork"></div>
                 
             </div>
+            <Reviewcard></Reviewcard>
         </div>
     )
 }
