+++ src/Reviewcard.jsx
@@ -21,7 +21,7 @@ const ReviewCard = ({ selectedCourse }) => {
     const savedVotes = JSON.parse(localStorage.getItem('userVotesMap') || '{}');
     setUserVotes(savedVotes);
     
-    const q = query(collection(db, "reviews", courseId, "items"), orderBy("createdAt", "desc"));
+    const q = query(collection(db, "reviews", courseId), orderBy("createdAt", "desc"));
     const unsubscribe = onSnapshot(q, (snapshot) => {
       setReviews(snapshot.docs.map(doc => ({ id: doc.id, ...doc.data() })));
     });
@@ -75,7 +75,7 @@ const ReviewCard = ({ selectedCourse }) => {
   const iconMap = { difficulty: DifficultyIcon, workload: WorkloadIcon, stress: StressIcon, enjoyment: EnjoymentIcon };
 
   const handleVote = async (reviewId, voteType) => {
-    const reviewRef = doc(db, "reviews", courseId, "items", reviewId);
+    const reviewRef = doc(db, "reviews", courseId, reviewId);
     const review = reviews.find(r => r.id === reviewId);
     const currentVote = userVotes[reviewId];
     let voteChange = currentVote === voteType ? (voteType === 'up' ? -1 : 1) : (currentVote ? (voteType === 'up' ? 2 : -2) : (voteType === 'up' ? 1 : -1));
@@ -91,7 +91,7 @@ const ReviewCard = ({ selectedCourse }) => {
 
   const submitReview = async () => {
     try {
-      await addDoc(collection(db, "reviews", courseId, "items"), {
+      await addDoc(collection(db, "reviews", courseId), {
         user: auth.currentUser?.displayName || 'temporaryguestuserthing',
         text: newReview.text,
         scores: { difficulty: Number(newReview.difficulty), workload: Number(newReview.workload), stress: Number(newReview.stress), enjoyment: Number(newReview.enjoyment) },
