+++ src/components/Homepage.jsx
@@ -1,5 +1,9 @@
 import { useState } from 'react'
 import {
+  Home,
+  Search,
+  SlidersHorizontal,
+  CircleUserRound,
   Calculator,
   BookOpen,
   FlaskConical,
@@ -25,7 +29,33 @@ const categories = [
   { label: 'All Classes', Icon: LayoutGrid },
 ]
 
-
+function SearchField({ className = '', value, onChange, onSubmit }) {
+  return (
+    <form
+      onSubmit={(e) => {
+        e.preventDefault()
+        onSubmit()
+      }}
+      className={`flex items-center gap-3 rounded-full bg-sif-search px-5 py-3 ${className}`}
+    >
+      <Search className="size-5 text-sif-green shrink-0" strokeWidth={2.5} />
+      <input
+        type="search"
+        placeholder="Search classes, teachers..."
+        value={value}
+        onChange={(e) => onChange(e.target.value)}
+        className="font-ui flex-1 bg-transparent text-base text-sif-green placeholder:text-sif-green/60 outline-none"
+      />
+      <button
+        type="submit"
+        aria-label="Search"
+        className="shrink-0 cursor-pointer rounded-full p-1 text-sif-green hover:bg-black/5"
+      >
+        <SlidersHorizontal className="size-5" strokeWidth={2.5} />
+      </button>
+    </form>
+  )
+}
 
 export default function Homepage({ toggleOverlay, goToCourseList }) {
   const [query, setQuery] = useState('')
@@ -40,6 +70,49 @@ export default function Homepage({ toggleOverlay, goToCourseList }) {
 
   return (
     <div className="font-ui min-h-svh bg-white text-sif-green">
+      <header className="sticky top-0 z-10 bg-sif-green">
+        <div className="mx-auto flex max-w-[1512px] items-center gap-3 px-4 py-3 sm:gap-4 sm:px-6 sm:py-4">
+          <a
+            href="#"
+            aria-label="Home"
+            className="grid size-11 shrink-0 place-items-center rounded-full text-white hover:bg-white/10 sm:size-12"
+          >
+            <Home className="size-6 sm:size-7" strokeWidth={2} />
+          </a>
+
+          <div className="hidden flex-1 md:block">
+            <SearchField
+              className="mx-auto max-w-2xl"
+              value={query}
+              onChange={setQuery}
+              onSubmit={submitSearch}
+            />
+          </div>
+
+          <div className="ml-auto flex items-center gap-2 sm:gap-3">
+            <button
+              type="button"
+              className="cursor-pointer rounded-[11px] bg-sif-green-accent px-4 py-2 text-sm font-bold text-white hover:brightness-110 sm:px-5 sm:py-2.5 sm:text-base"
+            >
+              Log in
+            </button>
+            <a
+              href="#"
+              className="hidden font-bold text-white hover:underline sm:inline"
+            >
+              Sign up
+            </a>
+            <button onClick={() => toggleOverlay()}
+              type="button"
+              aria-label="Account"
+              className="grid size-10 shrink-0 place-items-center rounded-full bg-white/10 text-white hover:bg-white/20 sm:size-12"
+            >
+              <CircleUserRound className="size-7 sm:size-8" strokeWidth={1.75} />
+            </button>
+          </div>
+        </div>
+      </header>
+
       <main className="mx-auto w-full max-w-[1512px] px-4 sm:px-6 lg:px-8">
         <section className="flex flex-col items-center pt-12 pb-10 text-center sm:pt-16 lg:pt-24">
           <h1 className="font-heading text-6xl font-bold tracking-tight text-sif-green sm:text-7xl lg:text-[90px]">
@@ -50,7 +123,12 @@ export default function Homepage({ toggleOverlay, goToCourseList }) {
             Also, review classes and see the opinions of your peers.
           </p>
 
-          
+          <SearchField
+            className="mt-8 w-full max-w-3xl md:hidden"
+            value={query}
+            onChange={setQuery}
+            onSubmit={submitSearch}
+          />
         </section>
 
         <section className="pt-6 pb-20 sm:pt-10">
